Format currency amounts in notifications
