+++ frontend/model/contracts/group.js
@@ -595,7 +595,8 @@ sbp('chelonia/defineContract', {
                 groupID: contractID,
                 creatorID: innerSigningContractID,
                 paymentHash: data.paymentHash,
-                amount: getters.withGroupCurrency(payment.data.amount)
+                amount: payment.data.amount,
+                currency: getters.groupSettings.mincomeCurrency
               })
             }
           }
