+++ frontend/views/containers/payments/PaymentRowSent.vue
@@ -55,7 +55,7 @@ import PaymentRow from './payment-row/PaymentRow.vue'
 import PaymentActionsMenu from './payment-row/PaymentActionsMenu.vue'
 import PaymentNotReceivedTooltip from './payment-row/PaymentNotReceivedTooltip.vue'
 import PaymentsMixin from '@containers/payments/PaymentsMixin.js'
-import withGroupCurrency from '@view-utils/withGroupCurrency.js'
+import { withGroupCurrency } from '@view-utils/misc.js'
 
 export default ({
   name: 'PaymentRowSent',
