+++ frontend/views/utils/misc.js
@@ -1,5 +1,25 @@
 'use strict'
+
+import sbp from '@sbp/sbp'
+import { DECIMALS_MAX } from '@model/contracts/shared/currencies.js'
+import { L } from '@common/common.js'
 import VueRouter from 'vue-router'
+
+// Used to avoid creating a new NumberFormat object every time an amount is formatted.
+// Note: if the user locale ever changes at runtime, cached currency formats won't update until page reload.
+const currencyFormatsByCode = Object.create(null)
+// $FlowIgnore[prop-missing]
+const supportedCurrencies = new Set(Intl.supportedValuesOf('currency'))
+
+// https://en.wikipedia.org/wiki/List_of_cryptocurrencies
+const symbolsByCode = {
+  'ADA': '₳',
+  'BTC': '₿',
+  'ETH': 'Ξ',
+  'LTC': 'Ł',
+  'XDG': 'Ð',
+  'XNO': 'Ӿ'
+}
 
 export function logExceptNavigationDuplicated (err: Object) {
   err.name !== 'NavigationDuplicated' && console.error(err)
@@ -75,4 +95,36 @@ export function validateURL (url: string, acceptPathOnly: boolean = false): Obje
   }
 
   return response
+}
+
+export function withCurrency (code: string, amount: number): string {
+  if (!currencyFormatsByCode[code]) {
+    currencyFormatsByCode[code] = new Intl.NumberFormat(
+      // $FlowIgnore[incompatible-call]
+      navigator.languages ?? navigator.language,
+      {
+        style: 'currency',
+        currency: code,
+        // If the currency is unsupported, then it's likely a cryptocurrency,
+        // in which case we have to set the number of decimal places explicitly.
+        maximumFractionDigits: supportedCurrencies.has(code) ? undefined : DECIMALS_MAX,
+        // Don't show fraction digits *if* they are all zero.
+        trailingZeroDisplay: 'stripIfInteger'
+      }
+    )
+  }
+  return symbolsByCode[code]
+    ? currencyFormatsByCode[code].format(amount).replace(code, symbolsByCode[code])
+    : currencyFormatsByCode[code].format(amount)
+}
+
+export function withGroupCurrency (amount: number): string {
+  // Group currency code.
+  const code = sbp('state/vuex/getters').groupSettings?.mincomeCurrency
+  // https://github.com/okTurtles/group-income/pull/2923#discussion_r2456246025
+  if (code === undefined) {
+    console.error('missing currency')
+    return L('ERROR: missing currency')
+  }
+  return withCurrency(code, amount)
 }
